+++ commands.h
@@ -4,14 +4,14 @@
 
 char * getCommand();
 int isNumber(char * str);
-int getAvgWordCount(int lineCounter, int * wordCounter);
+double getAvgWordCount(int lineCounter, int * wordCounter);
 void commandInputLoop(int k, int lineCounter, char ** lines, int * wordCounter, TrieNode * trie);
 void df(TrieNode * trie);
 int getDocumentFrequency(char * word, TrieNode * node);
 void printDocumentFrequency(CharList * word, TrieNode * node);
 void tf(TrieNode * trie);
 int getTermFrequency(int id, char * word, TrieNode * node);
-void search(int k, int * wordCounter, int avgWordCount, int lineCounter, TrieNode * trie);
-void loadTermsIntoTree(AVLTree ** tree, int * docCounter, int * wordCounter, int avgWordCount, int lineCounter, TrieNode * trie);
-double getScore(int tf, int wordCount, int avgWordCount, int lineCounter, int docFreq);
+void search(int k, int * wordCounter, double avgWordCount, int lineCounter, TrieNode * trie);
+void loadTermsIntoTree(AVLTree ** tree, int * docCounter, int * wordCounter, double avgWordCount, int lineCounter, TrieNode * trie);
+double getScore(int tf, int wordCount, double avgWordCount, int lineCounter, int docFreq);
 double getIDF(int lineCounter, int docFreq);
